PythonProject3/requirements.txt
Flask==3.0.0
google-generativeai==0.3.1
python-dotenv==1.0.0
selenium==4.15.2
webdriver-manager==4.0.1
Flask-Mail==0.9.1
itsdangerous==2.1.2
bleach==6.1.0
tenacity==8.2.3
reportlab==4.0.7
openpyxl==3.1.2
pytest==7.4.3
Werkzeug==3.0.1
PyPDF2==3.0.1
python-docx==1.1.0
youtube-transcript-api==0.6.2
beautifulsoup4==4.12.2
requests==2.31.0